requirements.txt
streamlit
pandass